requirements.txt
python-docx==0.8.11
Authlib~=1.3.0
python-dotenv~=0.21.0
flask~=3.0.0
future~=0.18.3
bs4~=0.0.1
beautifulsoup4~=4.11.1
requests~=2.28.1
splinter~=0.18.1
selenium~=4.4.3
pytube~=15.0.0
opencv-python~=4.7.0.72
ultralytics~=8.0.55
chromadb~=0.5.0
sentence_transformers
pymongo[srv]